get appinfos

+++ src/main/java/com/mbappe/newsql/user/persistence/ddl/AppInfoDO.java
@@ -28,7 +28,7 @@ public class AppInfoDO implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer groupId;
+    private Long groupId;
 
     /**
      *
@@ -112,7 +112,7 @@ public class AppInfoDO implements Serializable {
      *
      * @mbg.generated
      */
-    public Integer getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
@@ -124,7 +124,7 @@ public class AppInfoDO implements Serializable {
      *
      * @mbg.generated
      */
-    public void setGroupId(Integer groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
